Update to Besu 23.4.4

+++ acceptance-tests/dsl/src/main/java/org/hyperledger/besu/tests/acceptance/dsl/node/ThreadBesuNodeRunner.java
@@ -14,73 +14,76 @@
  */
 package org.hyperledger.besu.tests.acceptance.dsl.node;
 
-import static org.hyperledger.besu.cli.config.NetworkName.DEV;
 import static org.hyperledger.besu.controller.BesuController.DATABASE_PATH;
 
 import org.hyperledger.besu.Runner;
 import org.hyperledger.besu.RunnerBuilder;
 import org.hyperledger.besu.cli.config.EthNetworkConfig;
+import org.hyperledger.besu.cli.config.NetworkName;
+import org.hyperledger.besu.config.GenesisConfigFile;
+import org.hyperledger.besu.consensus.qbft.pki.PkiBlockCreationConfigurationProvider;
 import org.hyperledger.besu.controller.BesuController;
 import org.hyperledger.besu.controller.BesuControllerBuilder;
-import org.hyperledger.besu.crypto.KeyPairSecurityModule;
 import org.hyperledger.besu.crypto.KeyPairUtil;
-import org.hyperledger.besu.crypto.NodeKey;
+import org.hyperledger.besu.cryptoservices.KeyPairSecurityModule;
+import org.hyperledger.besu.cryptoservices.NodeKey;
+import org.hyperledger.besu.ethereum.GasLimitCalculator;
 import org.hyperledger.besu.ethereum.api.graphql.GraphQLConfiguration;
-import org.hyperledger.besu.ethereum.blockcreation.GasLimitCalculator;
 import org.hyperledger.besu.ethereum.eth.EthProtocolConfiguration;
 import org.hyperledger.besu.ethereum.eth.sync.SynchronizerConfiguration;
+import org.hyperledger.besu.ethereum.eth.transactions.ImmutableTransactionPoolConfiguration;
 import org.hyperledger.besu.ethereum.eth.transactions.TransactionPoolConfiguration;
-import org.hyperledger.besu.ethereum.p2p.peers.EnodeURL;
-import org.hyperledger.besu.ethereum.permissioning.PermissioningConfiguration;
+import org.hyperledger.besu.ethereum.p2p.peers.EnodeURLImpl;
 import org.hyperledger.besu.ethereum.storage.keyvalue.KeyValueStorageProvider;
 import org.hyperledger.besu.ethereum.storage.keyvalue.KeyValueStorageProviderBuilder;
+import org.hyperledger.besu.evm.internal.EvmConfiguration;
 import org.hyperledger.besu.metrics.MetricsSystemFactory;
 import org.hyperledger.besu.metrics.ObservableMetricsSystem;
+import org.hyperledger.besu.plugin.data.EnodeURL;
 import org.hyperledger.besu.plugin.services.BesuConfiguration;
 import org.hyperledger.besu.plugin.services.BesuEvents;
 import org.hyperledger.besu.plugin.services.PicoCLIOptions;
 import org.hyperledger.besu.plugin.services.SecurityModuleService;
 import org.hyperledger.besu.plugin.services.StorageService;
+import org.hyperledger.besu.plugin.services.TransactionSelectionService;
 import org.hyperledger.besu.plugin.services.storage.rocksdb.RocksDBPlugin;
+import org.hyperledger.besu.plugin.services.txselection.TransactionSelectorFactory;
 import org.hyperledger.besu.services.BesuConfigurationImpl;
 import org.hyperledger.besu.services.BesuEventsImpl;
 import org.hyperledger.besu.services.BesuPluginContextImpl;
+import org.hyperledger.besu.services.PermissioningServiceImpl;
 import org.hyperledger.besu.services.PicoCLIOptionsImpl;
+import org.hyperledger.besu.services.RpcEndpointServiceImpl;
 import org.hyperledger.besu.services.SecurityModuleServiceImpl;
 import org.hyperledger.besu.services.StorageServiceImpl;
+import org.hyperledger.besu.services.TransactionSelectionServiceImpl;
 
-import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.io.PrintStream;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.time.Clock;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
-import javax.annotation.Nonnull;
 
+import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.vertx.core.Vertx;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.ThreadContext;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.slf4j.MDC;
 import picocli.CommandLine;
 import picocli.CommandLine.Model.CommandSpec;
 
 public class ThreadBesuNodeRunner implements BesuNodeRunner {
 
-  private static final Logger LOG = LogManager.getLogger();
+  private static final Logger LOG = LoggerFactory.getLogger(ThreadBesuNodeRunner.class);
   private final Map<String, Runner> besuRunners = new HashMap<>();
 
   private final Map<Node, BesuPluginContextImpl> besuPluginContextMap = new ConcurrentHashMap<>();
-  private final PrintStream originalOut = System.out;
-  private final PrintStream originalErr = System.err;
-  private final ByteArrayOutputStream consoleContents = new ByteArrayOutputStream();
 
   private BesuPluginContextImpl buildPluginContext(
       final BesuNode node,
@@ -92,14 +95,22 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
     besuPluginContext.addService(StorageService.class, storageService);
     besuPluginContext.addService(SecurityModuleService.class, securityModuleService);
     besuPluginContext.addService(PicoCLIOptions.class, new PicoCLIOptionsImpl(commandLine));
+    besuPluginContext.addService(
+        TransactionSelectionService.class, new TransactionSelectionServiceImpl());
 
-    final Path pluginsPath = node.homeDirectory().resolve("plugins");
-    final File pluginsDirFile = pluginsPath.toFile();
-    if (!pluginsDirFile.isDirectory()) {
-      pluginsDirFile.mkdirs();
-      pluginsDirFile.deleteOnExit();
+    final Path pluginsPath;
+    final String pluginDir = System.getProperty("besu.plugins.dir");
+    if (pluginDir == null || pluginDir.isEmpty()) {
+      pluginsPath = node.homeDirectory().resolve("plugins");
+      final File pluginsDirFile = pluginsPath.toFile();
+      if (!pluginsDirFile.isDirectory()) {
+        pluginsDirFile.mkdirs();
+        pluginsDirFile.deleteOnExit();
+      }
+      System.setProperty("besu.plugins.dir", pluginsPath.toString());
+    } else {
+      pluginsPath = Path.of(pluginDir);
     }
-    System.setProperty("besu.plugins.dir", pluginsPath.toString());
     besuPluginContext.registerPlugins(pluginsPath);
 
     commandLine.parseArgs(node.getConfiguration().getExtraCLIOptions().toArray(new String[0]));
@@ -115,10 +126,10 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
   @Override
   public void startNode(final BesuNode node) {
 
-    if (ThreadContext.containsKey("node")) {
-      LOG.error("ThreadContext node is already set to {}", ThreadContext.get("node"));
+    if (MDC.get("node") != null) {
+      LOG.error("ThreadContext node is already set to {}", MDC.get("node"));
     }
-    ThreadContext.put("node", node.getName());
+    MDC.put("node", node.getName());
 
     if (!node.getRunCommand().isEmpty()) {
       throw new UnsupportedOperationException("commands are not supported with thread runner");
@@ -136,19 +147,25 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
                 buildPluginContext(
                     node, storageService, securityModuleService, commonPluginConfiguration));
 
+    GlobalOpenTelemetry.resetForTest();
     final ObservableMetricsSystem metricsSystem =
         MetricsSystemFactory.create(node.getMetricsConfiguration());
     final List<EnodeURL> bootnodes =
         node.getConfiguration().getBootnodes().stream()
-            .map(EnodeURL::fromURI)
+            .map(EnodeURLImpl::fromURI)
             .collect(Collectors.toList());
+    final NetworkName network = node.getNetwork() == null ? NetworkName.DEV : node.getNetwork();
     final EthNetworkConfig.Builder networkConfigBuilder =
-        new EthNetworkConfig.Builder(EthNetworkConfig.getNetworkConfig(DEV))
+        new EthNetworkConfig.Builder(EthNetworkConfig.getNetworkConfig(network))
             .setBootNodes(bootnodes);
     node.getConfiguration().getGenesisConfig().ifPresent(networkConfigBuilder::setGenesisConfig);
     final EthNetworkConfig ethNetworkConfig = networkConfigBuilder.build();
+    final SynchronizerConfiguration synchronizerConfiguration =
+        new SynchronizerConfiguration.Builder().build();
     final BesuControllerBuilder builder =
-        new BesuController.Builder().fromEthNetworkConfig(ethNetworkConfig);
+        new BesuController.Builder()
+            .fromEthNetworkConfig(
+                ethNetworkConfig, Collections.emptyMap(), synchronizerConfiguration.getSyncMode());
 
     final KeyValueStorageProvider storageProvider =
         new KeyValueStorageProviderBuilder()
@@ -157,29 +174,82 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
             .withMetricsSystem(metricsSystem)
             .build();
 
-    final BesuController besuController =
-        builder
-            .synchronizerConfiguration(new SynchronizerConfiguration.Builder().build())
-            .dataDirectory(node.homeDirectory())
-            .miningParameters(node.getMiningParameters())
-            .privacyParameters(node.getPrivacyParameters())
-            .nodeKey(new NodeKey(new KeyPairSecurityModule(KeyPairUtil.loadKeyPair(dataDir))))
-            .metricsSystem(metricsSystem)
-            .transactionPoolConfiguration(TransactionPoolConfiguration.DEFAULT)
-            .ethProtocolConfiguration(EthProtocolConfiguration.defaultConfig())
-            .clock(Clock.systemUTC())
-            .isRevertReasonEnabled(node.isRevertReasonEnabled())
-            .storageProvider(storageProvider)
-            .gasLimitCalculator(GasLimitCalculator.constant())
+    final TransactionPoolConfiguration txPoolConfig =
+        ImmutableTransactionPoolConfiguration.builder()
+            .strictTransactionReplayProtectionEnabled(node.isStrictTxReplayProtectionEnabled())
             .build();
 
-    final RunnerBuilder runnerBuilder = new RunnerBuilder();
-    if (node.getPermissioningConfiguration().isPresent()) {
-      final PermissioningConfiguration permissioningConfiguration =
-          node.getPermissioningConfiguration().get();
+    final int maxPeers = 25;
 
-      runnerBuilder.permissioningConfiguration(permissioningConfiguration);
-    }
+    final Optional<TransactionSelectorFactory> transactionSelectorFactory =
+        getTransactionSelectorFactory(besuPluginContext);
+
+    builder
+        .synchronizerConfiguration(new SynchronizerConfiguration.Builder().build())
+        .dataDirectory(node.homeDirectory())
+        .miningParameters(node.getMiningParameters())
+        .privacyParameters(node.getPrivacyParameters())
+        .nodeKey(new NodeKey(new KeyPairSecurityModule(KeyPairUtil.loadKeyPair(dataDir))))
+        .metricsSystem(metricsSystem)
+        .transactionPoolConfiguration(txPoolConfig)
+        .ethProtocolConfiguration(EthProtocolConfiguration.defaultConfig())
+        .clock(Clock.systemUTC())
+        .isRevertReasonEnabled(node.isRevertReasonEnabled())
+        .storageProvider(storageProvider)
+        .gasLimitCalculator(GasLimitCalculator.constant())
+        .pkiBlockCreationConfiguration(
+            node.getPkiKeyStoreConfiguration()
+                .map(pkiConfig -> new PkiBlockCreationConfigurationProvider().load(pkiConfig)))
+        .evmConfiguration(EvmConfiguration.DEFAULT)
+        .maxPeers(maxPeers)
+        .lowerBoundPeers(maxPeers)
+        .maxRemotelyInitiatedPeers(15)
+        .networkConfiguration(node.getNetworkingConfiguration())
+        .randomPeerPriority(false)
+        .transactionSelectorFactory(transactionSelectorFactory);
+
+    node.getGenesisConfig()
+        .map(GenesisConfigFile::fromConfig)
+        .ifPresent(builder::genesisConfigFile);
+
+    final BesuController besuController = builder.build();
+
+    final RunnerBuilder runnerBuilder = new RunnerBuilder();
+    runnerBuilder.permissioningConfiguration(node.getPermissioningConfiguration());
+
+    runnerBuilder
+        .vertx(Vertx.vertx())
+        .besuController(besuController)
+        .ethNetworkConfig(ethNetworkConfig)
+        .discovery(node.isDiscoveryEnabled())
+        .p2pAdvertisedHost(node.getHostName())
+        .p2pListenPort(0)
+        .networkingConfiguration(node.getNetworkingConfiguration())
+        .jsonRpcConfiguration(node.jsonRpcConfiguration())
+        .webSocketConfiguration(node.webSocketConfiguration())
+        .jsonRpcIpcConfiguration(node.jsonRpcIpcConfiguration())
+        .dataDir(node.homeDirectory())
+        .metricsSystem(metricsSystem)
+        .permissioningService(new PermissioningServiceImpl())
+        .metricsConfiguration(node.getMetricsConfiguration())
+        .p2pEnabled(node.isP2pEnabled())
+        .p2pTLSConfiguration(node.getTLSConfiguration())
+        .graphQLConfiguration(GraphQLConfiguration.createDefault())
+        .staticNodes(
+            node.getStaticNodes().stream()
+                .map(EnodeURLImpl::fromString)
+                .collect(Collectors.toList()))
+        .besuPluginContext(new BesuPluginContextImpl())
+        .autoLogBloomCaching(false)
+        .storageProvider(storageProvider)
+        .rpcEndpointService(new RpcEndpointServiceImpl());
+    node.engineRpcConfiguration().ifPresent(runnerBuilder::engineJsonRpcConfiguration);
+
+    final Runner runner = runnerBuilder.build();
+
+    besuPluginContext.beforeExternalServices();
+
+    runner.startExternalServices();
 
     besuPluginContext.addService(
         BesuEvents.class,
@@ -190,37 +260,10 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
             besuController.getSyncState()));
     besuPluginContext.startPlugins();
 
-    final Runner runner =
-        runnerBuilder
-            .vertx(Vertx.vertx())
-            .besuController(besuController)
-            .ethNetworkConfig(ethNetworkConfig)
-            .discovery(node.isDiscoveryEnabled())
-            .p2pAdvertisedHost(node.getHostName())
-            .p2pListenPort(0)
-            .maxPeers(25)
-            .networkingConfiguration(node.getNetworkingConfiguration())
-            .jsonRpcConfiguration(node.jsonRpcConfiguration())
-            .webSocketConfiguration(node.webSocketConfiguration())
-            .dataDir(node.homeDirectory())
-            .metricsSystem(metricsSystem)
-            .metricsConfiguration(node.getMetricsConfiguration())
-            .p2pEnabled(node.isP2pEnabled())
-            .graphQLConfiguration(GraphQLConfiguration.createDefault())
-            .staticNodes(
-                node.getStaticNodes().stream()
-                    .map(EnodeURL::fromString)
-                    .collect(Collectors.toList()))
-            .besuPluginContext(new BesuPluginContextImpl())
-            .autoLogBloomCaching(false)
-            .storageProvider(storageProvider)
-            .build();
-
-    System.setOut(null);
-    runner.start();
+    runner.startEthereumMainLoop();
 
     besuRunners.put(node.getName(), runner);
-    ThreadContext.remove("node");
+    MDC.remove("node");
   }
 
   @Override
@@ -263,44 +306,20 @@ public class ThreadBesuNodeRunner implements BesuNodeRunner {
     }
   }
 
-  static class SplittingStream extends OutputStream {
-    final OutputStream one;
-    final OutputStream two;
-
-    SplittingStream(final OutputStream one, final OutputStream two) {
-      this.one = one;
-      this.two = two;
-    }
-
-    @Override
-    public void write(final int b) throws IOException {
-      one.write(b);
-      two.write(b);
-    }
-
-    @Override
-    public void write(@Nonnull final byte[] b) throws IOException {
-      one.write(b);
-      two.write(b);
-    }
-
-    @Override
-    public void write(@Nonnull final byte[] b, final int off, final int len) throws IOException {
-      one.write(b, off, len);
-      two.write(b, off, len);
-    }
-  }
-
   @Override
   public void startConsoleCapture() {
-    System.setOut(new PrintStream(new SplittingStream(consoleContents, originalOut)));
-    System.setErr(new PrintStream(new SplittingStream(consoleContents, originalErr)));
+    throw new RuntimeException("Console contents can only be captured in process execution");
   }
 
   @Override
   public String getConsoleContents() {
-    System.setOut(originalOut);
-    System.setErr(originalErr);
-    return consoleContents.toString(StandardCharsets.UTF_8);
+    throw new RuntimeException("Console contents can only be captured in process execution");
+  }
+
+  private Optional<TransactionSelectorFactory> getTransactionSelectorFactory(
+      final BesuPluginContextImpl besuPluginContext) {
+    final Optional<TransactionSelectionService> txSelectionService =
+        besuPluginContext.getService(TransactionSelectionService.class);
+    return txSelectionService.isPresent() ? txSelectionService.get().get() : Optional.empty();
   }
 }
